% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameChange.R
\name{aliasChange}
\alias{aliasChange}
\title{aliasChange}
\usage{
aliasChange(platdata, aliases)
}
\arguments{
\item{platdata}{see nameChange}

\item{aliases}{see nameChange}

\item{platcols}{see nameChange}
}
\value{
see nameChange
}
\description{
This function performs the alias based gene renaming; called by nameChange
}
