% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameChange.R
\name{blastChange}
\alias{blastChange}
\title{blastChange}
\usage{
blastChange(platdata, platform, transcripts, match, secmatch)
}
\arguments{
\item{platdata}{see nameChange}

\item{platform}{curated data from nameChange}

\item{transcripts}{see nameChange}

\item{match}{see nameChange}

\item{secmatch}{see nameChange}
}
\value{
see nameChange
}
\description{
This function is...elaborate. It requires blast be installed and be in your
PATH. This function takes your platform data file and your annotated
transcriptome file and actually creates a blast database, searches all probes
against it, and identifies hits that are unique and maximal (depending on
your match and secondary match parameters.) It's called from nameChange.
}
