% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRCgeneration.R
\name{CRCgeneration}
\alias{CRCgeneration}
\title{CRCgeneration}
\usage{
CRCgeneration(stages)
}
\arguments{
\item{stages}{This is a single column matrix; row names are sample names, and
the column contains the calculated IDC location for each sample (in hpi)}
}
\value{
This function returns a 3 column matrix; the first column contains
  the provided location in hpi, and the additional columns contain the
  calculated x and y covariates, respectively.
}
\description{
CRCgeneration
}
