% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftimpute.R
\name{fftimpute}
\alias{fftimpute}
\title{fftimpute}
\usage{
fftimpute(expdata, times)
}
\arguments{
\item{expdata}{An expression matrix with genes as rows and samples as columns}

\item{times}{A matrix with row names as samples and one column containing sampling times}
}
\value{
This function returns an expression matrix containing both the measured and imputed timepoints
}
\description{
This function imputes timepoints using the Fourier Transform; strongly recommend using another option here, as this is INCREDIBLY specific to a 2 hr timecourse over 56 hr
}
