% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPhase.R
\name{FFTphase}
\alias{FFTphase}
\title{FFTphase}
\usage{
FFTphase(expdata, times)
}
\arguments{
\item{expdata}{see peakPhase}

\item{times}{see peakPhase}
}
\value{
see peakPhase
}
\description{
This function calculates peak phase for each gene in a time course using the
Fourier Transform.
}
