% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameChange.R
\name{flattenAlias}
\alias{flattenAlias}
\title{flattenAlias}
\usage{
flattenAlias(aliasfilename)
}
\arguments{
\item{aliasfilename}{File name of alias file obtained from PlasmoDB.}
}
\value{
Data frame with two columns; first column is old identifiers, and
  second column is their mapping to the most recent name (per the alias file)
}
\description{
The alias file from PlasmoDB is unwieldy and awful. Here is a function for
flattening it for easier find and replace. Called in aliasChange, you don't
need to use this for anything.
}
