% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullCurate.R
\name{fullCurate}
\alias{fullCurate}
\title{fullCurate}
\usage{
fullCurate(
  GSE,
  platcols = c(1, 2),
  method = "blast",
  aliases = "current",
  transcripts = "current",
  match = 130,
  secmatch = 60,
  pct = 0.8,
  pullmeta = T
)
}
\arguments{
\item{GSE}{This is the GSE accession number.}

\item{platcols}{These are the columns necessary for the nameChange function;
see nameChange for details.}

\item{method}{These are the available methods for name changes; see
nameChange for details.}

\item{aliases}{IF method is 'aliases', list the alias file name or 'current'
for a PlasmoDB download. See nameChange for details.}

\item{transcripts}{If method is 'transcripts', this is the transcriptome
fasta file, or if 'current' it will download the most recent file.}

\item{match}{see curateExpData.}

\item{secmatch}{See curateExpData.}

\item{pct}{See curateExpData.}

\item{pullmeta}{See curateExpData.}
}
\value{
See curateExpData.
}
\description{
Are you feeling really, spectacularly lazy? This is the function for you. If
you know the GEO accession number of your expression set and the columns
necessary for nameChange, this function will do the entire curation for you
in one call.
}
