% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogTransform.R
\name{getAverages}
\alias{getAverages}
\title{GetAverages
This function will get the average expression of all genes across a reference time course. Note: the reference time course should be evenly samples across a single IDC.}
\usage{
getAverages(reftc)
}
\arguments{
\item{reftc}{This is a reference time course to average; this must be a numeric matrix with genes as rows and samples as columns.}
}
\value{
This function returns a matrix with row names as gene names and a single column containing the average expression across the reference time course.
}
\description{
GetAverages
This function will get the average expression of all genes across a reference time course. Note: the reference time course should be evenly samples across a single IDC.
}
