% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lSmooth.R
\name{lSmooth}
\alias{lSmooth}
\title{lSmooth}
\usage{
lSmooth(expdata, time, smoothvalue = 0.3)
}
\arguments{
\item{expdata}{Expression data matrix, with columns as samples and rows as genes}

\item{time}{matrix with one column; names are samples and column contains sample time}

\item{smoothvalue}{loess smoothing value; default is 0.3}
}
\value{

}
\description{
This function will loess smooth a time course, first demonstrated on P. fal time courses by Bozdech et al 2003
}
