% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameChange.R
\name{nameChange}
\alias{nameChange}
\title{nameChange}
\usage{
nameChange(
  platform,
  method = "swap",
  aliases = NA,
  transcripts = NA,
  match = 130,
  secmatch = 60
)
}
\arguments{
\item{platform}{This is the filename for the platform from GEO. A simple
cut/paste in to a txt file will work here (it's tab delimited by default).}

\item{method}{There are 3 methods here; "swap" just takes the new gene name
from the platform file (least accurate), "alias" will find new names based
on the current PlasmoDB assembly file (not bad), "blast" will take a
provided Annotated Transcripts file from PlasmoDB, blast each probe against
it, and only keep probes that meet the bitscore criteria (default is 130
for perfect hit and 60 for minimal secondary hits).}

\item{aliases}{This is the flat alias file created by the function
"flattenAlias"; only necessary if selected method is "alias".}

\item{transcripts}{The fasta file containing the annotated transcripts; this
can be obtained from PlasmoDB.}

\item{match}{The bitscore for a perfect primary match; only necessary if
"blast" is selected, default is 130.}

\item{secmatch}{The maxmium bitscore for secondary probe alignments; only
necessary if "blast is selected, default is 60.}
}
\value{

}
\description{
This function takes a platform file from GEO (a cut and paste from the full
table view, really), and identifies the new genome assembly name for each
probe using one of a few methods. This function is a wrapper for subfunctions
that carry out each method.
}
