% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPhase.R
\name{peakPhase}
\alias{peakPhase}
\title{peakPhase}
\usage{
peakPhase(expdata, times, method = "poly")
}
\arguments{
\item{expdata}{This must be an expression matrix, with rows as genes and
columns as samples. Sample names (column names) must match sample names in
the "times" parameter.}

\item{times}{This must be a matrix with one column; the column is actual
sampling time, and the row names are the sample names.}

\item{method}{This allows the user to choose the method for peak time
detemination; current options are "poly"(default), "sin", "spline", and
"FFT". Note, FFT will only work on evenly sampled time courses (or evenly
imputed time courses).}
}
\value{
This function returns a data frame with a single column; the row
  names are the gene names, and the column contains the time at which the
  gene is calculated to express maximally.
}
\description{
peakPhase
}
