% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPhase.R
\name{polyPhase}
\alias{polyPhase}
\title{polyPhase}
\usage{
polyPhase(expdata, times)
}
\arguments{
\item{expdata}{see peakPhase}

\item{times}{see peakPhase}
}
\value{
see peakPhase
}
\description{
This is a sub function that calculates peak expression by 6th order
polynomial fit. Accessed through peakPhase.
}
