% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remLowCoverage.R
\name{remLowCoverage}
\alias{remLowCoverage}
\title{remLowCoverage}
\usage{
remLowCoverage(expdata, mincov, faillist = F)
}
\arguments{
\item{expdata}{This is a matrix containing expression values' rows must be genes and columns must be samples.}

\item{mincov}{This is the minimum percent coverage; probes must appear in at least this percentage of samples to pass QC.}

\item{faillist}{Default is FALSE; if TRUE, function returns a list containing both the curated expression matrix and a list of genes that failed QC.}
}
\value{
This function returns an expression matrix with the genes that do not pass QC removed.
}
\description{
Removes any gene from the expression dataset that does not have values in a minimum percentage of samples.
}
