% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPhase.R
\name{sinPhase}
\alias{sinPhase}
\title{sinPhase}
\usage{
sinPhase(expdata, times)
}
\arguments{
\item{expdata}{see peakPhase}

\item{times}{see peakPhase}
}
\value{
see peakPhase
}
\description{
This is a sub function that calculates peak expression by sin fitting.
Accessed through peakPhase.
}
