% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineCourse.R
\name{splineCourse}
\alias{splineCourse}
\title{splineCourse}
\usage{
splineCourse(expdata, times, fulltimes = 1:48)
}
\arguments{
\item{expdata}{Actual expression data; a matrix with genes as rows and time
course samples as columns. In order from earliest to latest.}

\item{times}{The actual sample times for the expression matrix.}

\item{fulltimes}{A vector containing the times desired; for example, an hourly
time course over a single IDC would be 1:48.}
}
\value{
Returns an expression matrix expanded with additional columns to contain samples at the desired times.
}
\description{
This function takes an existing time course and uses spline interpolation to
create an imputed time course with the desired resolution.
}
