% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPhase.R
\name{splinePhase}
\alias{splinePhase}
\title{splinePhase}
\usage{
splinePhase(expdata, times)
}
\arguments{
\item{expdata}{see peakPhase}

\item{times}{see peakPhase}
}
\value{
see peakPhase
}
\description{
This function fits a spline to the expression data and determines the time at
which each gene is maximally expressed. Called by peakPhase.
}
