% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagingByTranscription.R
\name{stagingByTranscription}
\alias{stagingByTranscription}
\title{stagingByTranscription}
\usage{
stagingByTranscription(
  samples,
  reference,
  subset = NA,
  cortable = F,
  heatmap = T,
  start = 1
)
}
\arguments{
\item{samples}{A matrix of expression samples, with rows as genes and columns
as samples. This is the set that will be tested.}

\item{reference}{A reference expression time course, with rows as genes and
columns as samples.}

\item{subset}{The default is to use all genes in common between the two sets;
if a subset is desired, provide this as vector of gene names in the subset.}

\item{cortable}{If the full matrix of correlation values is desired, set to
TRUE; default is FALSE.}

\item{heatmap}{A simple correlation heatmap will be generated as default. Set
to false to skip.}

\item{start}{What hour the reference course starts at- default is 1hpi}
}
\value{
Default return is a matrix with one column; rows are samples and
  values are the time point of maximum correlation. The function will also
  plot a simple heatmap by default. If cortable is set to TRUE, the function
  will return a list; the first entry is the simple time point matrix, and
  the second entry is the full correlation matrix.
}
\description{
This function compares each whole transcriptome sample to every time point in
the provided reference time course, and identifies the best time point fit.
Correlation is determined by the Pearson method.
}
