% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameChange.R
\name{swapChange}
\alias{swapChange}
\title{swapChange}
\usage{
swapChange(platdata)
}
\arguments{
\item{platdata}{see nameChange}
}
\value{
see nameChange
}
\description{
This function simply pulls the new ORF and probe ID and creates a renaming
file from that; I swear this wasn't there before, but now that we have the
new GeneIDs listed might as well build a function to rename based on them.
This function is called by nameChange.
}
