% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcImpute.R
\name{tcImpute}
\alias{tcImpute}
\title{tcImpute}
\usage{
tcImpute(expdata, neighbors)
}
\arguments{
\item{expdata}{This is an expression matrix with rows as genes and columns as samples in a time course. Time points missing data should be inserted as columns in the correct place containing NAs.}

\item{neighbors}{This is the number of neighbors used in determining value.}
}
\value{
This fucntion returns the expression matrix with blank columns filled with imputed values.
}
\description{
This function calls the bnstruct package "knn.impute" function to perform k nearest neighbor imputation on a time course to fill in missing data.
}
