% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curateExpData.R
\name{curateExpdata}
\alias{curateExpdata}
\title{curateExpdata}
\usage{
curateExpdata(expname, namelist, local = T, pct = 0.75)
}
\arguments{
\item{expname}{This is the name of your data set; if local is T, it is a GEO
file (any type containing exp data). If local is F, this can be a GEO
accession number.}

\item{namelist}{This is a name list file as generated by nameChange.}

\item{local}{This is a logical variable; if T, the function looks for a local
file. If F, the function uses the expname as a GEO accession number and
attempts to download the data.}

\item{pct}{This is the cutoff for gene data; if a gene has expression values
for less than this percentage of samples, it is excluded.}
}
\value{
Returns an expression frame containing genes as rows and samples as
  columns.
}
\description{
This function will take the name lists generated by nameChange and apply them
to an expression data set. The function will process locally downloaded GEO
expression files, or even just accept a GEO accession number and do the work
for you.
}
\examples{

}
